cell: change width type from usize to u8
Cell is an often used structure, this should come with some decent
memory savings since there are cells stored twice (Screen and
InternalScreen).

+++ src/Cell.zig
@@ -23,7 +23,7 @@ pub const Character = struct {
     /// will measure the same width. This can be ensure by using the gwidth method
     /// included in libvaxis. If width is 0, libvaxis will measure the glyph at
     /// render time
-    width: usize = 1,
+    width: u8 = 1,
 };
 
 pub const CursorShape = enum {
